% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_classes.R
\name{get_col_classes}
\alias{get_col_classes}
\title{get_col_classes}
\usage{
get_col_classes(data)
}
\arguments{
\item{country}{Country in which SAM site is located}
}
\value{
Janes equipment data.
}
\description{
Get col classes of
}
