% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes.R
\name{get_janes}
\alias{get_janes}
\title{get_janes}
\usage{
get_janes(
  country = NULL,
  parallel = FALSE,
  endpoint = c("inventories", "equipment", "orbats", "bases", "airports",
    "countryrisks", "companies", "events", "equipmentrelationships", "samsites",
    "ewsites", "nuclearsites", "satelliteImages", "marketforecasts", "defenceprogrammes",
    "references", "news")
)
}
\arguments{
\item{country}{Country filter for equipment}

\item{parallel}{Whether to use parallel processing}

\item{endpoint}{What endpoint to pull data from}
}
\value{
Janes data
}
\description{
General interface to Janes data
}
