% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_airports.R
\name{get_janes_airports}
\alias{get_janes_airports}
\title{get_janes_airports}
\usage{
get_janes_airports(country = NULL)
}
\arguments{
\item{country}{Country in which airport is located}
}
\value{
Janes airports data.
}
\description{
Pulls Janes airports data.
}
