% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_companies.R
\name{get_janes_companies}
\alias{get_janes_companies}
\title{get_janes_companies}
\usage{
get_janes_companies(country = NULL, query = NULL)
}
\arguments{
\item{country}{Country in which base is located}

\item{query}{Search term for companies (i.e. company name)}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes company data
}
