% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_country_risks.R
\name{get_janes_country_risks}
\alias{get_janes_country_risks}
\title{get_janes_country_risks}
\usage{
get_janes_country_risks(country = NULL)
}
\arguments{
\item{country}{Country in which base is located}
}
\value{
Janes country risk data
}
\description{
Pulls Janes country risk data
}
