% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_data_json.R
\name{get_janes_data_json}
\alias{get_janes_data_json}
\title{get_janes_data_json}
\usage{
get_janes_data_json(x)
}
\arguments{
\item{x}{For iteration on URLs}
}
\value{
Janes data. Used in all "get_janes_x" functions.
}
\description{
Pulls Janes data for all data endpoints. Helper function.
}
