% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_equipment_relationships.R
\name{get_janes_equipment_relationships}
\alias{get_janes_equipment_relationships}
\title{get_janes_equipment_relationships}
\usage{
get_janes_equipment_relationships(country = NULL, query = NULL)
}
\arguments{
\item{country}{Country filter for equipment}

\item{query}{Query filter for equipment}
}
\value{
Janes equipment hierarchical links.
}
\description{
Pulls Janes equipment relationships (hierarchical links).
}
