% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_events.R
\name{get_janes_events}
\alias{get_janes_events}
\title{get_janes_events}
\usage{
get_janes_events(
  country = NULL,
  query = NULL,
  post_date = NULL,
  start_date = NULL,
  event_type = c("Terrorism and Insurgency", "Intelligence Events")
)
}
\arguments{
\item{country}{Event country}

\item{query}{Search term for events}

\item{post_date}{Event post date}

\item{start_date}{Event start date}

\item{event_type}{Either "Terrorism and Insurgency" or "Intelligence Events"}
}
\value{
Janes events data
}
\description{
Pulls Janes events data
}
