% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_events.R
\name{get_janes_events_jtic}
\alias{get_janes_events_jtic}
\title{get_janes_events_jtic}
\usage{
get_janes_events_jtic(
  country = NULL,
  query = NULL,
  post_date = NULL,
  start_date = NULL
)
}
\arguments{
\item{country}{Event country}

\item{query}{Search term for events}

\item{post_date}{Event post date}

\item{start_date}{Event start date}
}
\value{
Janes events data
}
\description{
Pulls Janes events data
}
