% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_ew.R
\name{get_janes_ew}
\alias{get_janes_ew}
\title{get_janes_ew}
\usage{
get_janes_ew(country = NULL)
}
\arguments{
\item{country}{Country in which EW site is located}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes EW sites data
}
