% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_info.R
\name{get_janes_info}
\alias{get_janes_info}
\title{get_janes_info}
\usage{
get_janes_info(
  country = NULL,
  branch = NULL,
  type = NULL,
  endpoint = c("inventories", "equipment", "orbats", "bases", "airports",
    "defenceprogrammes"),
  x
)
}
\arguments{
\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Of base}

\item{endpoint}{One of 6 options currently}

\item{x}{For iteration on URLs}
}
\value{
Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes data info for all data endpoints. Helper function.
}
