% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_info_programs.R
\name{get_janes_info_programs}
\alias{get_janes_info_programs}
\title{get_janes_info_programs}
\usage{
get_janes_info_programs(x, operator_country = NULL)
}
\arguments{
\item{x}{For iteration on URLs}

\item{operator_country}{Country filter for news}
}
\value{
Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes data info for all data endpoints. Helper function.
}
