% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_json.R
\name{get_janes_json}
\alias{get_janes_json}
\title{get_janes_json}
\usage{
get_janes_json(
  country = NULL,
  branch = NULL,
  type = NULL,
  operator_force = NULL,
  query = NULL,
  environment = NULL,
  endpoint = c("inventories", "equipment", "orbats", "bases", "airports",
    "countryrisks", "companies", "events", "equipmentrelationships", "references",
    "samsites", "ewsites")
)
}
\arguments{
\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Depends on endpoint}

\item{operator_force}{Operator force}

\item{query}{Search term}

\item{environment}{Of search, i.e. "Air"}

\item{endpoint}{One of 6 options currently}
}
\value{
Janes data in JSON format.
}
\description{
Pulls Janes data as a JSON file for all endpoints.
}
