% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_markets_forecast.R
\name{get_janes_markets_forecast}
\alias{get_janes_markets_forecast}
\title{get_janes_markets_forecast}
\usage{
get_janes_markets_forecast(
  end_user_country = NULL,
  query = NULL,
  market = NULL,
  subsystems = FALSE
)
}
\arguments{
\item{end_user_country}{Country in which base is located}

\item{query}{Search term}

\item{market}{JMF market to query}

\item{subsystems}{Include subsystems data. Logical.}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes Markets Forecast data.
}
