% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_news.R
\name{get_janes_news}
\alias{get_janes_news}
\title{get_janes_news}
\usage{
get_janes_news(country = NULL, query = NULL)
}
\arguments{
\item{country}{Country filter for news}

\item{query}{Keyword search for news}
}
\value{
Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes news articles page range for a given search. Helper function.
}
