% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_news_text.R
\name{get_janes_news_text}
\alias{get_janes_news_text}
\title{get_janes_news_text}
\usage{
get_janes_news_text(x)
}
\arguments{
\item{x}{urls to pull from}
}
\value{
Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes news articles page range for a given search. Helper function.
}
