% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_nuclear_sites.R
\name{get_janes_nuclear_sites}
\alias{get_janes_nuclear_sites}
\title{get_janes_nuclear_sites}
\usage{
get_janes_nuclear_sites(country = NULL)
}
\arguments{
\item{country}{Country in which base is located}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes nuclear site data
}
