% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_orbats.R
\name{get_janes_orbats}
\alias{get_janes_orbats}
\title{get_janes_orbats}
\usage{
get_janes_orbats(country = NULL, branch = NULL)
}
\arguments{
\item{country}{Country in which base is located}

\item{branch}{Military branch (Air Force, Army, Navy)}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes order of battle (ORBAT) data.
}
