% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_programs.R
\name{get_janes_programs}
\alias{get_janes_programs}
\title{get_janes_programs}
\usage{
get_janes_programs(operator_country = NULL, histories = FALSE)
}
\arguments{
\item{operator_country}{Country to pull programs from}

\item{histories}{Pull only program revision history data. Logical.}
}
\value{
Janes defense programs data.
}
\description{
Pulls Janes defense programs data
}
