% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_samsites.R
\name{get_janes_sam}
\alias{get_janes_sam}
\title{get_janes_sam}
\usage{
get_janes_sam(country = NULL)
}
\arguments{
\item{country}{Country in which SAM site is located}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes SAM sites data
}
