% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_sat_images.R
\name{get_janes_sat_images}
\alias{get_janes_sat_images}
\title{get_janes_sat_images}
\usage{
get_janes_sat_images(country = NULL, query = NULL, pics = FALSE)
}
\arguments{
\item{country}{Country of satellite image target}

\item{query}{Keyword search}

\item{pics}{If TRUE, returns satellite images as JPG files. If FALSE, returns metadata}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes satellite images
}
