% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_xml.R
\name{get_janes_xml}
\alias{get_janes_xml}
\title{get_janes_xml}
\usage{
get_janes_xml(
  country = NULL,
  branch = NULL,
  type = NULL,
  operator_force = NULL,
  query = NULL,
  environment = NULL,
  endpoint = NULL,
  jdet_only = FALSE,
  docs = TRUE,
  pics = FALSE,
  sat = FALSE
)
}
\arguments{
\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Depends on endpoint}

\item{operator_force}{Operator force}

\item{query}{Search term}

\item{environment}{Of search, i.e. "Air"}

\item{endpoint}{One of either "news" or "references"}
}
\value{
Janes data in xml format.
}
\description{
Pulls Janes data as a xml file for all endpoints.
}
