% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_range_programs.R
\name{get_page_range_programs}
\alias{get_page_range_programs}
\title{get_page_range_programs}
\usage{
get_page_range_programs(operator_country = NULL)
}
\arguments{
\item{operator_country}{Country filter for news}
}
\value{
Janes page ranges for a given search.
}
\description{
Pulls Janes page ranges for defense programs. Helper function
}
