% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_janes_key.R
\name{save_janes_key}
\alias{save_janes_key}
\title{save_janes_key}
\usage{
save_janes_key(key)
}
\arguments{
\item{key}{Janes API key}
}
\value{
Saves Janes API key to environment.
}
\description{
Saves Janes key to environment
}
