% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_all.R
\name{unnest_all}
\alias{unnest_all}
\title{unnest_all}
\usage{
unnest_all(data, endpoint)
}
\arguments{
\item{data}{Country filter for equipment}

\item{variable}{Query filter for equipment}
}
\value{
Unnests_wider if variable is in dataset, doesn't if not.
}
\description{
Helper function to unnest all list columns. If no list_cols, cleans names
}
