% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_bases.R
\name{get_janes_bases}
\alias{get_janes_bases}
\title{get_janes_bases}
\usage{
get_janes_bases(country = NULL, type = NULL)
}
\arguments{
\item{country}{Country in which base is located}

\item{type}{Type of base}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes bases data.
}
