% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_equipment.R
\name{get_janes_equipment}
\alias{get_janes_equipment}
\title{get_janes_equipment}
\usage{
get_janes_equipment(country = NULL, query = NULL)
}
\arguments{
\item{country}{Country filter for news}
}
\value{
Janes equipment data.
}
\description{
Pulls Janes equipment.
}
