% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_events.R
\name{get_janes_events}
\alias{get_janes_events}
\title{get_janes_events}
\usage{
get_janes_events(country = NULL, query = NULL)
}
\arguments{
\item{country}{Event country}

\item{query}{Search term for events}
}
\value{
Janes events data
}
\description{
Pulls Janes events data
}
