% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_info.R
\name{get_janes_info}
\alias{get_janes_info}
\title{get_janes_info}
\usage{
get_janes_info(
  x,
  country = NULL,
  branch = NULL,
  type = NULL,
  operator_force = NULL,
  query = NULL,
  environment = NULL,
  post_date = NULL,
  endpoint = c("inventories", "equipment", "orbats", "bases", "airports",
    "countryrisks", "companies", "events", "equipmentrelationships", "references",
    "samsites", "ewsites", "satelliteImages")
)
}
\arguments{
\item{x}{For iteration on URLs}

\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Filter for type - depends on endpoint (i.e. "Platforms" or "Air")}

\item{operator_force}{= Operator force}

\item{query}{Search term}

\item{environment}{Filter for environment of equipment (i.e. "Air")}

\item{endpoint}{One of 6 options currently}
}
\value{
Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes data info for all data endpoints. Helper function.
}
