% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_info.R
\name{get_janes_info}
\alias{get_janes_info}
\title{get_janes_info}
\usage{
get_janes_info(
  x,
  country = NULL,
  branch = NULL,
  type = NULL,
  operator_force = NULL,
  query = NULL,
  environment = NULL,
  post_date = NULL,
  start_date = NULL,
  event_type = NULL,
  market = NULL,
  end_user_country = NULL,
  overall_family = NULL,
  endpoint = c("inventories", "equipment", "orbats", "news", "bases", "airports",
    "countryrisks", "companies", "events", "equipmentrelationships", "references",
    "samsites", "ewsites", "satelliteImages", "marketforecasts", "nuclearsites")
)
}
\arguments{
\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Depends on endpoint}

\item{operator_force}{Operator force}

\item{query}{Query}

\item{environment}{Of search, i.e. "Air"}

\item{post_date}{Event post date}

\item{start_date}{Event start date}

\item{event_type}{Event type - JTIC or Intel Event}

\item{market}{Markets Forecast market}

\item{end_user_country}{JMF end user country}

\item{overall_family}{Overall equipment family}

\item{endpoint}{One of 6 options currently}
}
\value{
Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes data info for all data endpoints. Helper function.
}
