% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_equipment.R,
%   R/get_janes_inventories.R
\name{get_janes_inventories}
\alias{get_janes_inventories}
\title{get_janes_equipment}
\usage{
get_janes_inventories(country = NULL)

get_janes_inventories(country = NULL)
}
\arguments{
\item{country}{Country filter for news}
}
\value{
Janes equipment data.

Janes equipment inventories.
}
\description{
Pulls Janes equipment.

Pulls Janes equipment inventories.
}
