% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_inventories.R
\name{get_janes_inventories}
\alias{get_janes_inventories}
\title{get_janes_inventories}
\usage{
get_janes_inventories(country = NULL, operator_force = NULL)
}
\arguments{
\item{country}{Country filter}

\item{operator_force}{Operator force}
}
\value{
Janes equipment inventories.
}
\description{
Pulls Janes equipment inventories.
}
