% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janes_news.R, R/get_news_page_range.R
\name{get_news_page_range}
\alias{get_news_page_range}
\title{get_janes_news}
\usage{
get_news_page_range(country = NULL, query = NULL)

get_news_page_range(country = NULL, query = NULL)
}
\arguments{
\item{country}{Country filter for news}

\item{query}{Keyword search for news}
}
\value{
Returns Janes news articles in dataframe related to search.

Helper function to return Janes news article page range related to search.
}
\description{
Pulls Janes news articles for a given search.

Pulls Janes news articles page range for a given search. Helper function.
}
