% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_range.R, R/get_page_range_programs.R
\name{get_page_range}
\alias{get_page_range}
\title{get_page_range}
\usage{
get_page_range(country = NULL)

get_page_range(country = NULL)
}
\arguments{
\item{country}{Country filter for news}

\item{query}{Keyword search for news}

\item{branch}{Military branch}

\item{type}{Of base}

\item{endpoint}{One of 6 options currently}

\item{operator_country}{Country filter for news}
}
\value{
Janes page ranges for a given search.

Janes page ranges for a given search.
}
\description{
Pulls Janes page ranges for all data endpoints. Helper function

Pulls Janes page ranges for defense programs. Helper function
}
