% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_range.R
\name{get_page_range}
\alias{get_page_range}
\title{get_page_range}
\usage{
get_page_range(
  country = NULL,
  branch = NULL,
  type = NULL,
  operator_force = NULL,
  query = NULL,
  environment = NULL,
  post_date = NULL,
  event_type = NULL,
  endpoint = c("inventories", "equipment", "orbats", "bases", "airports",
    "countryrisks", "companies", "events", "equipmentrelationships", "references",
    "samsites", "ewsites", "satelliteImages")
)
}
\arguments{
\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Depends on endpoint}

\item{operator_force}{Operator force}

\item{query}{Search term}

\item{environment}{Of search, i.e. "Air"}

\item{endpoint}{One of 6 options currently}
}
\value{
Janes page ranges for a given search.
}
\description{
Pulls Janes page ranges for all data endpoints. Helper function
}
