% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_range.R
\name{get_page_range}
\alias{get_page_range}
\title{get_page_range}
\usage{
get_page_range(
  country = NULL,
  branch = NULL,
  type = NULL,
  endpoint = c("inventories", "equipment", "orbats", "bases", "airports")
)
}
\arguments{
\item{country}{Country filter for news}

\item{branch}{Military branch}

\item{type}{Of base}

\item{endpoint}{One of 6 options currently}
}
\value{
Janes page ranges for a given search.
}
\description{
Pulls Janes page ranges for all data endpoints. Helper function
}
